Zoom
Added differing camera zooms

+++ Source/Weapon_System/WeaponStatLibrary.h
@@ -183,7 +183,7 @@ public:
 		float Recoil;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-		int Zoom;
+		float Zoom;
 };
 
 USTRUCT(BlueprintType)
